Add new files and update existing documents
This commit includes the addition of new binary data to the OpenDocument files `entwickeln Weg.odt`, `was brauchen wir.odt`, and `Neue in UI.odt`, as well as modifications to `game over.png` and `main view.png`. The updates enhance the content and structure of the documents, including new metadata, configuration settings, and styles. Manifest files have also been updated to ensure proper referencing of resources and document integrity.